% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df}
\alias{df}
\title{MTCars, formatted}
\format{
A data frame with 32 rows and 11 variables:
\describe{
\item{mpg}{miles per gallon}
\item{cyl}{number of cylinders}
\item{disp}{displacement (cubic inches)}
\item{hp}{horsepower}
\item{drat}{rear axle ratio}
\item{wt}{weight (thousands of pounds)}
\item{qsec}{quarter mile time (seconds)}
\item{vs}{engine shape: 0 = V-shape, 1 = Straight}
\item{am}{transmission: 0 = automatic, 1 = manual}
\item{gear}{number of gears: 3 = 3, 4 = 4, 5 = 5}
\item{carb}{number of carburetors: 1 = 1, 2 = 2, 3 = 3, 4 = 4, 6 = 6, 8 = 8}
}
}
\source{
R base dataset
}
\usage{
df
}
\description{
The mtcars dataset has been formatted to include factors for engine shape, transmission, number of gears, and number of carburetors.
}
\keyword{datasets}
