% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rto.R
\name{rto}
\alias{rto}
\title{ReadyTableOne: A Quick Table 1 of Descriptive Statistics}
\usage{
rto(x, stratify, nonpara = FALSE)
}
\arguments{
\item{x}{data frame containing numeric and factor variables}

\item{stratify}{variable by which to stratify (e.g., treatment)}

\item{nonpara}{use non-parametric estimates (i.e., median, IQR)}
}
\value{
table of descriptive statistics
}
\description{
Create a publication-ready table of descriptive statistics stratified by binary variable.
}
\details{
This package will prepare a high-quality table of descriptive statistics for a pre-formatted dataframe containing numeric and factor variables.
}
\examples{
 
rto(df, stratify = "vs")
}
\author{
Jake Cofferen
}
