% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rto.R
\name{rto}
\alias{rto}
\title{add two numbers}
\usage{
rto(x, stratify = y, nonpara = FALSE)
}
\arguments{
\item{x}{data frame containing numeric and factor variables}

\item{stratify}{variable by which to stratify (e.g., treatment)}

\item{nonpara}{use non-parametric estimates (i.e., median, IQR)}
}
\value{
table of descriptive statistics
}
\description{
Create a publication-ready table (Description)
}
\details{
This package will prepare a high-quality table of descriptive statistics for a pre-formatted dataframe containing numeric and factor variables. (Details)
}
\author{
Jake Cofferen
}
